Move all subcommand code to sub_comamnd module
