+++ src/sub_command/mod.rs
@@ -1,0 +1,9 @@
+// This file is part of cfnts.
+// Copyright (c) 2019, Cloudflare. All rights reserved.
+// See LICENSE for licensing information.
+
+//! Subcommand collections.
+
+pub mod client;
+pub mod ke_server;
+pub mod ntp_server;
